# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "DER-EMS API",
        "contact": {},
        "license": {}
    },
    "basePath": "/api",
    "paths": {
...
        "/{gwid}/devices/accumulated-power-state": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get power state by token, gateway UUID, resolution, startTime and endTime",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analysis"
                ],
                "summary": "Show daily/monthly accumulated power state of Load/Solar/Battery/Grid",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Input user's access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Gateway UUID",
                        "name": "gwid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "date-time",
                        "example": "UTC time in ISO-8601",
                        "name": "endTime",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "day",
                            "month"
                        ],
                        "type": "string",
                        "name": "resolution",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "date-time",
                        "example": "UTC time in ISO-8601",
                        "name": "startTime",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/services.AccumulatedPowerStateResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
        "/{gwid}/devices/energy-distribution-info": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get energy distribution by token, gateway UUID, startTime and endTime",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analysis"
                ],
                "summary": "Show the distribution of energy sources and distinations",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Input user's access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Gateway UUID",
                        "name": "gwid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "date-time",
                        "example": "UTC time in ISO-8601",
                        "name": "endTime",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "date-time",
                        "example": "UTC time in ISO-8601",
                        "name": "startTime",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/services.EnergyDistributionInfoResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
        "/{gwid}/devices/power-self-supply-rate": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get power self supply rate by token, gateway UUID, resolution, startTime and endTime",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analysis"
                ],
                "summary": "Show daily/monthly power self supply rate",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Input user's access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Gateway UUID",
                        "name": "gwid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "date-time",
                        "example": "UTC time in ISO-8601",
                        "name": "endTime",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "day",
                            "month"
                        ],
                        "type": "string",
                        "name": "resolution",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "date-time",
                        "example": "UTC time in ISO-8601",
                        "name": "startTime",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/services.PowerSelfSupplyRateResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/{gwid}/devices/power-state": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get power state by token, gateway UUID, resolution, startTime and endTime",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analysis"
                ],
                "summary": "Show today's hourly power state of Load/Solar/Battery/Grid",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Input user's access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Gateway UUID",
                        "name": "gwid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "date-time",
                        "example": "UTC time in ISO-8601",
                        "name": "endTime",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "hour",
                            "5minute"
                        ],
                        "type": "string",
                        "name": "resolution",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "date-time",
                        "example": "UTC time in ISO-8601",
                        "name": "startTime",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/services.PowerStateResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "app.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "services.AccumulatedPowerStateResponse": {
            "type": "object",
            "properties": {
                "batteryLifetimeEnergyACDiffs": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "gridLifetimeEnergyACDiffs": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "loadConsumedLifetimeEnergyACDiffs": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "pvProducedLifetimeEnergyACDiffs": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "timestamps": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "services.EnergyDistributionInfoResponse": {
            "type": "object",
            "properties": {
                "allConsumedLifetimeEnergyACDiff": {
                    "type": "number"
                },
                "allProducedLifetimeEnergyACDiff": {
                    "type": "number"
                },
                "batteryConsumedEnergyPercentAC": {
                    "type": "number"
                },
                "batteryConsumedLifetimeEnergyACDiff": {
                    "type": "number"
                },
                "batteryProducedEnergyPercentAC": {
                    "type": "number"
                },
                "batteryProducedLifetimeEnergyACDiff": {
                    "type": "number"
                },
                "gridConsumedEnergyPercentAC": {
                    "type": "number"
                },
                "gridConsumedLifetimeEnergyACDiff": {
                    "type": "number"
                },
                "gridProducedEnergyPercentAC": {
                    "type": "number"
                },
                "gridProducedLifetimeEnergyACDiff": {
                    "type": "number"
                },
                "loadConsumedEnergyPercentAC": {
                    "type": "number"
                },
                "loadConsumedLifetimeEnergyACDiff": {
                    "type": "number"
                },
                "pvProducedEnergyPercentAC": {
                    "type": "number"
                },
                "pvProducedLifetimeEnergyACDiff": {
                    "type": "number"
                }
            }
        },
...
        "services.PowerSelfSupplyRateResponse": {
            "type": "object",
            "properties": {
                "loadSelfConsumedEnergyPercentACs": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "timestamps": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "services.PowerStateResponse": {
            "type": "object",
            "properties": {
                "batteryAveragePowerACs": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "gridAveragePowerACs": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "loadAveragePowerACs": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "pvAveragePowerACs": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "timestamps": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
    }
}